Release version 5.4.6 Branch: main Deployed by: Clearpay Plugins[Bot]

+++ Plugin/Model/Order/Payment/Operations/ProcessInvoiceOperation.php
@@ -99,7 +99,7 @@ class ProcessInvoiceOperation
             // prepare parent transaction and its amount
             $paidWorkaround = 0;
             if (!$invoice->wasPayCalled()) {
-                $paidWorkaround = (double)$amountToCapture;
+                $paidWorkaround = (float)$amountToCapture;
             }
             if ($payment->isCaptureFinal($paidWorkaround)) {
                 $payment->setShouldCloseParentTransaction(true);
